重构预览逻辑并引入ImageData类
在 `Preview.cs` 中：
* 引入了 `MSDFAtlasGenerator.Models` 命名空间。
* 将 `RgbaProperty` 替换为 `ImageDataProperty`，并在属性更改时调用 `UpdatePreview` 方法。
* 将 `Rgba` 属性替换为 `ImageData` 属性。
* 修改了 `OnRender` 方法以使用 `ImageBrush` 绘制图像。
* 添加了 `UpdatePreview` 方法，用于更新 `previewImage` 并调用 `InvalidateVisual`。

在 `Atlas.cs` 中：
* 将 `width` 和 `height` 属性的类型从 `double` 改为 `int`。

在 `GeneratorViewModel.cs` 中：
* 引入了 `MSDFAtlasGenerator.Models` 命名空间。
* 添加了 `previewData` 属性。
* 修改了 `Preview` 方法，在生成预览时设置 `PreviewData`。

在 `GeneratorPage.xaml` 中：
* 修改了 `Preview` 控件，绑定 `ImageData` 属性到 `PreviewData`。

在 `ImageData.cs` 中：
* 新增了 `ImageData` 类，继承自 `ObservableObject`。
* 添加了 `width`、`height` 和 `data` 属性，并在构造函数中进行初始化。

+++ MSDFAtlasGenerator/Models/Atlas.cs
@@ -18,10 +18,10 @@ public partial class Atlas : ObservableObject
     private double size;
 
     [ObservableProperty]
-    private double width;
+    private int width;
 
     [ObservableProperty]
-    private double height;
+    private int height;
 
     [ObservableProperty]
     private YDirection yOrigin;
